Fix unit test on Linux
The file timestamp resolution on Linux/Mono is 1s whereas on Windows
it is 1ms. This change increases the time to wait between the
creation of two files to 1s. In order to make the unit tests faster
we do this only for the one test where the order of the files (based
on the timestamp) matters.

NOTE: This is only a hack. If we're relying on the timestamps for
sorting the files in the real app and it is possible that two files
get created within a few milliseconds then we have a real problem
that this hack hides.