POC: Get lexicon entries from a hardcoded project
This is closer to what we'll actually be doing: iterating through all
the fields, and processing them in different ways depending on what they
are (multitexts, optionlists, and so on). Here we just use the "lexicon"
field and iterate through its values, but we also see how the "config"
property of the project can give us all the field names.

Also note the difference between FirstOrDefaultAsync().Result and
ToListAsync().Result.FirstOrDefault() in the two alternate ways to get
the project. Both work, it's just a matter of which is easier to read.