Add MongoDB connector and mapper classes
This allows us to query MongoDB and get back actual classes in our query
results, not just untyped BsonDocuments. That way we can write code like
project.Config.Fields["lexicon"].Type, not project["config"]["fields"],
etc., and have Intellisense auto-complete field names for us.

This won't be perfect, as (for example) when we get Fields["lexicon"] we
will have to cast it to an LfConfigMultiText instance, and if the
LanguageForge data model ever changes, that cast will fail at runtime.
A better solution might involve figuring out the casts dynamically with
reflection, but I'm not sure the gain would be worth the cost (reflection
is slow, for example). Worth investigating further.

This commit also includes some hardcoded testing code in Program.cs,
which should eventually be moved to proper unit tests.