MongoDB mappings for lexicon entries
One part of this API is still ugly: MultiText fields are a dictionary
that maps writing system IDs to LfStringField instances (which have just
one property named Value). I'd like to map them to dicts of strings to
strings instead, and bypass the LfStringField mapping. But for now, this
works even if it's ugly, so it's worth committing.

Also, UpdateFdoFromMongoDbAction has a small amount of testing code.