Test with TestLangProj instead of Sena 3
Sena 3 isn't importing into LanguageForge, but TestLangProj is.

Note that the TestLangProj version currently (as of 2015-11-06) checked
into fwrepo isn't fully up-to-date with what the code expects: the
version in the repo has <languageproject version="7000062">, but the FW
code is up to version 7000068. Since we're disabling data migrations,
we need a version of TestLangProj that's up to 7000068. Opening it in
FieldWorks, then copying the resulting project folder over to
fwdata/fw/DistFiles/ReleaseData is enough for now.