Elegant but not-working generic factory function
This would be a much more elegant way to remove the duplicated code in
the LfFooToFdoFoo functions, but unfortunately it won't work. It is
correct in principle, but the IFooFactory and IFooRepository classes in
FDO weren't actually registered with the service locator as generic
IRepository<T> and IFactory<T> services. So I'll have to do this in a
slightly less elegant way, which will be in the next commit.