WIP on Mongo -> FDO direction: use just 1 UOW
Since NonUndoableUnitOfWorkHelper doesn't support nested units of work,
use just one unit of work for the entire operation. That means that
either the entire LF->FDO update will succeed as a whole unit, or else
the entire thing will be rolled back and leave the FDO project intact.
That may or may not be desired behavior in every case once we go to
production, but it's a good default: it guarantees that the user won't
see half-successful updates with mixed data.

This code doesn't do much yet, but it's worth checking in since
switching to a single unit of work touches a lot of lines, so it's good
to have this commit as a base so that future diffs will be smaller.