Fix null reference errors with new Sense objects
ILexSense objects, and a few other classes, require an owner to be
passed into the factory's Create() method or they won't be set up
properly. (The problem we were seeing is that their Cache attribute
was null, producing failure later on.) Tweak GetOrCreateCmObjectByGuid()
to pass in an owner if the incoming class requires an owner.