WIP on Mongo->FDO direction: custom fields started
To get custom field data out of Mongo and put it into FDO objects
correctly, we need to store the GUIDs of some types of custom fields
(particularly ICmPossibility references). The FDO->Mongo code that I'd
previously written for custom fields didn't store the GUIDs. This is one
attempt at solving the problem, but it's quite ugly. I plan to refactor
this approach to be more elegant in the next commit. For now, though,
I'll commit this code since it's a step in the right direction.