WIP on FDO->Mongo direction: persist values to DB
CAUTION: This does NOT work yet. The values are currently being
persisted with extra _t fields that Mongo uses as discriminators to tell
what class was persisted. I believe this is because, for some reason,
Mongo is unable to read the data types from the object properties in our
BuildUpdate() function. But since a lot of work went into getting even
this far, I want to commit what I have so far before continuing to work
on BuildUpdate().