Persist values to Mongo without _t fields
This should fix the BuildUpdate() function. Turns out that Mongo's
UpdateBuilder.Set() function is generic, and figures out the "expected"
type of the field from the type of its second argument. Since I was
using PropertyInfo.GetValue(), which returns object, Set() was expecting
an object and getting something derived from object instead. As per the
https://groups.google.com/forum/#!topic/mongodb-user/QGctV4Hbipk thread,
this was causing _t fields to be persisted. The solution was to cast the
second argument of Set() explicitly to the correct type; then Mongo was
expecting that type and was happy to persist it.