Don't persist empty or null fields
If the LfLexEntry (or any other object) contains any null values or
empty lists, that just means that FieldWorks had no data to persist for
those cases. In which case, deleting the data in Mongo (if any) is
probably not the right thing to do.

OTOH, maybe deleting the Mongo data in such cases *will* turn out to be
the right thing to do, and we'll revert this commit after further
reflection. It's possible.