Reverse earlier decision: DO commit empty values
On further reading of the spec, we're only going to be pushing data into
MongoDB after a merge has been done in FDO land. So if any fields are
empty at the point of putting them into MongoDB, that's because they
*should* be empty -- any data in those fields that used to exist in
MongoDB should be deleted.

However, we still need the BuildUpdate function, since we don't want to
just overwrite the Mongo document with our own. There are some fields,
like the "id" field that LanguageForge uses internally, that we need to
preserve. So we do need to perform an update procedure.