First steps towards sub-updates
Since we need to NOT blow away the "id" field on senses & examples
(because LanguageForge uses it internally), we can't just do a simple
update of the entry -- because it would set the "senses" field to:

"senses": [{"gloss": "foo", ...}]

Note the lack of an "id" field in there! Instead, we need to do a
sub-update, like:

"$set": {"lexeme": "fu", "senses.$.gloss": "foo", ...}

This is the first step towards doing sub-updates.