Don't use sub-updates after all
After discussing the LfMerge spec and process with Ira, decided that we
don't need to do sub-updates after all. Each sense and example in a
lexicon entry will acquire a Guid field ($guid in PHP), which will be
useable to tell different senses apart, so the $id field that currently
exists will be okay to blow away. (The LF code will need to look for a
$guid field first, and use $id if $guid does not exist -- that will
allow projects without send/receive to continue telling senses and
examples apart, while projects *with* send/receive can round-trip to FDO).

The upshot of this is that LfMerge's updates to Mongo just became easy:
we don't have to deal with sub-updates, so the code works fine as-is.