Fix '"_csharpnull": true' showing up in Mongo
Mont of the time, when you give MongoDB's C# driver a null, it will be
serialized as null into BSON. But if the type of the field was BsonValue
or one of its derived classes (including BsonDocument), a null will be
serialized as '"_csharpnull": true'. This was showing up in just one
field, and the fix is to simply change the type of that field.