Add GOLDEtic.xml template and parsing scripts
The GOLDEtic.xml file comes from the FieldWorks source, where it is used
to set up the default catalog of grammatical categories (parts of speech)
available to the user in FieldWorks projects. We want to have a copy here
so that we can use it in matching parts of speech by either their GUIDs
or their English names.

The output from parsme.sh will be in a format suitable for pasting into
C# source code as four dictionaries: a forward and reverse dictionary
for both flat names-to-GUIDs and hierarchical names-to-GUIDs. We may not
use all of those in the end, but since they're easy to construct at the
same time, we might as well do so.