First steps towards processing parts of speech
We add a PartOfSpeechConverter class that looks up part-of-speech names
from the master list (parsed from GOLDEtic.xml) and converts them to
well-known GUIDs, so as to get the right PartOfSpeech objects from FDO
when words like "adjective" or "possessive pronoun" are found in LF data.

Next step: handling cases where the part of speech object is NOT found
in FDO, or where there is a part of speech with an "older" GUID. Those
cases are trickier, and will probably involve parsing the GOLDEtic.xml
file in C#. (We should also parse it only once, and cache the result).