Tweak GoldEticXmlParser API a little
We don't need to use an instance here, as there's nothing we need to
persist. This is essentially just a namespace for one function.

We also make the GoldEticItem.AddChild() member private, as it is
designed to be called only from the constructor and could cause
inconsistencies if called from outside code.