Create new PartOfSpeech objects if not found
If the part of speech from LanguageForge isn't found, look up its name
in the GOLDEtic.xml data for the user's interface writing system to get
the "official" GUID for that part of speech. Whether or not we get an
"official" GUID, create a new PartOfSpeech object in FDO if one doesn't
already exist.

This completes the part-of-speech handling, which should now be correct.