Create interface for LfMergeSettings
This is in preparation for switching to an IniParser-based settings
class. The first step is to create a settings interface, so that we can
eventually switch to an AutoFac-registered settings class without
disrupting any existing code.

One minor change to existing code needed to be made: the
Queue.CreateQueueDirectories() function uses LfMergeSettings.Current. It
should instead use a settings object passed in as a parameter; this will
allow dependency injection in the future. (For non-static methods, we'll
use constructor injection, but for static methods like this one,
constructor injection isn't suitable.)