Switch to using ILfMergeSettings interface
Code that referred directly to the LfMergeSettings class is now using
the interface instead. We have NOT yet switched over the uses of
LfMergeSettings.Current everywhere, but it is now defined as an
ILfMergeSettings object and everything still compiles, which means the
interface is not missing any members.

NOTE: The interface just got one new method, SaveSettings(). This is
only used in unit testing, and I'm not sure it should be on the
interface. I think the unit test that currently uses it should be split
into two different tests, one for each implementation. Then each test
can use the specific implementation's SaveSettings() method. But to
solve the problem quickly, I added SaveSettings() to the interface
instead. At some point it should be removed and both implementations
should be properly tested.