Simpler PartOfSpeechConverter.CreateFromCustomName
NOTE: This might not handle "A|b|c" cases, where A is an "official" PoS
from GOLDEtic.xml, but b and c are "custom" PoSes. I think what may
happen is that A will be created with the wrong GUID. Need to write a
unit test case for that scenario.