Better use of LINQ Zip() function
Don't use Zip() just for side-effects. Since all I was doing was adding
the result to a list, I should have just been returning that result from
the Zip() inner function and building the list one at a time that way.
(And besides, all I was doing with the list was enumerating it later in
a Select() call, so why not make it an IEnumerable in the first place?)