Add SIL.WritingSystems.dll dependency, and use it
This is needed to have access to the CurrentAnalysisWritingSystems list,
which is a list of CoreWritingSystemDefinition objects.

(CoreWritingSystemDefinition is derived from WritingSystemDefinition,
which is found in the new SIL.WritingSystems.dll code.)

Then we use that list to improve BestStringFromMultiText(), which now
searches through all the analysis writing systems in order, rather than
just looking for the default analysis writing system.