Add a copy of FW's TestLangProj to data/ folder
This will be useful in testing. This copy of TestLangProj has been
updated fwdata version 7000069, which is the most current version
as of 2016-01-18. By checking in a "clean" copy that LfMerge has not yet
touched, we'll be able to ensure that our tests can run repeatedly
against the same data.