Improve part of speech handling in Mongo->FDO
Newly-created LexSense objects won't have a MorphoSyntaxAnalysisRA
member yet, so we need to create one. The proper way to do that in FDO
is to use the SandboxGenericMSA class.

Also, set up derivational affix handling to, at some point, be able to
set both From and To parts of speech (once the LF model allows that).