Actually update Mongo and FDO in round-trip tests
The "MongoConnectionDoubleThatStoresData" thing is kind of ugly: why
should there be a difference between the two types of connections? We
should probably try to consolidate them.

Now the test is down to just one difference, and that might be simply
due to a different GUID in the sample data. We'll have to find out.