Stop leaving big test environments around
Now that we have a good way to compare "before" and "after" values in
FDO and/or Mongo, we don't really need to leave the temporary files
around for manual comparison. They're often large (the TestLangProj
project files total about 90 megabytes), so leaving hundreds of them
around in /tmp will quickly run you out of memory.