Add more profiling code
It seems the unit tests are slow is because creating an FDO cache is slow
(about 6 seconds or so for a project the size of TestLangProj). Copying
the TestLangProj to a new directory is quite fast. That's bad news for
the unit test suite, though, as we really do need to create a new FDO
cache for most (if not all) FDO-related tests.