Improve fix for failing test on CI machine
One of the ReceiveActionTests failed on Jenkins that didn't fail
on my machine because /var/lib/languageforge was writable on the
CI machine and so the test went through a different code path.
This also revealed that the test wasn't quite correct in it's
expectations. This change improves the test and accounts for both
possible failure reasons.