WIP on part of speech handling: almost done
Two unit tests are failing with real failures (due to how TestLangProj
has old values for grammar info), and need to be addressed. The
TestLangProj names and abbreviations for grammar should have been
copied into Mongo, and weren't.

One unit test is failing but isn't a real failure, because the old
MongoConnectionDouble class isn't adequate anymore. I need to remove
that class and replace it entirely with the new MongoConnection double
that keeps data (whose name also needs to be shorter).