Now use hierarchical abbreviations as PoS keys
Since some abbreviations in the TestLangProj data are duplicates (e.g.,
"rel" for "relator" and for "relative pronoun") use ORC-delimited
abbreviations as keys instead. Subtituting pipe characters for ORC for
the sake of readability, then you'd have "rel" for "relator" and
"Pro|rel" for "relative pronoun" in the TestLangProj data.