Fix some FDO->Mongo grammar bugs, plus unit tests
One test still failing, but the rest of these tests now pass. (Some of
them revealed bugs, like the fact that we were creating new part of
speech objects with the wrong name).

The failing test shows that we are creating *two* grammar entries in Fdo
when we should really only create *one*. Will find out why and fix in
next commit.