Previous unit test wasn't failing, but this one is
Previous failure was actually the code working as designed, where two
parts of speech would be created if a well-known PoS had a "parent" in
the GOLDEtic.xml hieraechy. Since we had "subordinating connector",
whose parent is "connector", in our test, it was properly creating two
entries.

However, creating a second test to verify that the parent's name and
abbreviation are correct... failed, because the parent's abbreviation is
wrong. If we look at PartOfSpeechMasterList.HierarchicalPosAbbrevs,
we'll see that the parent abbreviations are actually names. The parsing
code is wrong, and needs to be corrected.