Improve custom string field conversion
Now we'll actually look for the best writing system for the field
instead of just taking the first non-empty value.

Note that WritingSystemServices.ActualWs looked promising, but it bases
its search on the data found in the FDO object, and we need to search
the Mongo data, not the FDO object. So we have to more or less duplicate
the logic from ActualWs().