Update TestLangProj to FDO model version 7000069
Currently, LfMerge is building from the versions of libpalaso and FieldWorks 
that have the SIL.WritingSystems update.

This commit checks in the migrated TestLangProj that the unit tests use.
This new version of TestLangProj also includes 3 test images and 5 AudioVisual wav files.

TODO: Fix the Action tests that involve updating pictures.