Make code compatible with either FW 8 or 9
In most cases, the same code can be used in both FW 8 and 9 if you make
sure the use the relevant interfaces (ILgWritingSystem instead of
CoreWritingSystemDefinition, and ILgWritingSystemFactory instead of
WritingSystemManager). In one function, this doesn't work, and we use
a compiler #if...#endif directive. Compiling for FW 8 will require
defining FW8_COMPAT.