Copy FW's FixFwData project into our own repo
This is necessary so we can tweak its build process without breaking the
main FieldWorks build on Windows.

Note that to build FixFwData, you need the FixFwDataDll.dll file from a
FieldWorks build. Currently, we don't have a good way to automatically
download that file from Jenkins, so I've added a few lines to the
build/update-dependencies script to say "Please download this file
manually from the following URL". This is an ugly hack, and we'll want
something automated later on -- but the BuildDependencyTasks code (which
is what processes LfMerge.dep) currently doesn't have a way to handle
Jenkins servers. Once it does, we can remove that ugly hack.