WIP to preserve GUIDs in custom MultiPara fields
This is not quite working: LfStringField doesn't preserve GUIDs when
persisting to Mongo, so paragraphs end up doubled when you round-trip
(because the original paragraphs have their GUIDs stripped, so they look
like new ones). To fix this, we need to add a Guids field to
LfStringMultiText, which can be empty, and only preserve it if it's
non-empty. Doable, but annoying since LfStringField is used almost
everywhere yet only needs to know about Guids for one special case.