Enable failing SynchronizeAction test
SynchronizeAction_LFDataChangedLDDataDeleted_LFWins is currently
failing.  We think we may have some  Flexbridge/Chorus configuration issue
so enabling the test for Randy / Eberhard to help troubleshoot.

The scenario is LF modifies an entry, while LanguageDepot deletes the entry
We expect the modified entry to remain after the merge, but for some reason
the entry is deleted.