Get path to php source files from settings
To support building on arbitrary build agents it is necessary that
the files required for running tests can be included in the source
tree.

With this change the php source files are expected to live under
data/php, which can be easiest achieved by cloning/copying the
web-languageforge repo to data/php.