MongoConnectionDouble now clones data in and out
Previously, when you called GetLfLexEntries() on MongoConnectionDouble,
it would return the .Values attribute of its dictionary. Which meant
that every call to GetLfLexEntries() was returning the same list of
objects. If a test saved the original list, then made changes to some of
the LfLexEntry objects and called GetLfLexEntries() again expecting the
new list to be different, it would fail because the changes would have
been reflected in the "original" list. This commit fixes that problem.

Also, one of the tests was (accidentally) written with the assumption
that the GetLfLexEntries() result *should* update to reflect changes
made later by the test. That's not what we intended, so I've changed
that test.