Fix up testlangproj in our Git repo
It turns out that the issues that FixFwData is complaining about are
inherent in the testlangproj.fwdata file, not added by our code. So we
will run FixFwData once against testlangproj and commit the result, and
use the fixed-up version of the project in future testing.

We also did the same for the testlangproj-modified project used in the
synchronization tests.