Add testlangproj data WITHOUT extraneous bulk
For times when it becomes necessary to look at the testlangproj.fwdata
file in a text editor, it might be useful to have a copy of it WITHOUT
the bulky semantic domain data. Even bulkier is the many, many ChkRef
objects that contain lists of Scriptural terms that might need checking;
these, too, can be ignored for our purposes.

I renamed the file extension from .fwdata to .xml so that this
stripped-down file can't be loaded into FW. Because I didn't touch the
LexEntry, LexSense, etc. objects in any way, many of them contain GUID
references to semantic domain objects that are no longer present, so
this is no longer a valid .fwdata file. Thus the extension change.