Try to work around custom fields data bug in FDO
In FDO, when you set a value in a custom field, it registers a change
whether or not the value actually changed. This causes DateModified to
be updated, which in turn causes Mercurial (and Chorus) to see a change
in that LexEntry. We'll try to work around the bug by checking the old
value of the custom field before setting the new value, and only setting
the new value if it has really changed.