Try to fix Mercurial in .deb package
Some Mercurial modules like osutil are distributed as .so files, but the
Debian source-package build process defaults to silently strip .so files
(as well as many other files like *.o and .gitignore and .svn) from the
source package before creating the tarball. If you set explicit
tar-ignore options, though, it will not use its defaults, and will only
ignore the files specified. This should fix the Mercurial problems we'd
been having with the lfmerge .deb package.