Gulp script: give up on sequencing, just use shell
After several attempts at sequencing the gulp-rsync task with a gulp-ssh
task, all of which failed because the gulp-ssh task was started before
the gulp-rsync task was *actually* finished, I'm giving up and just
using the gulp-shell task instead. This essentially makes Gulp just a
big complicated wrapper around a two-line shell script, but it's not
worth spending ANOTHER three days on this script.