Modify lfmerge launcher to run from /usr/lib/lfmerge
Some parts of the system currently rely on the current working directory
to find files like FixFwData.exe. That has been fixed in a PR that's not
yet merged in, but to make the tests, etc., run on the dev server right
now, we need to ensure that we run from the /usr/lib/lfmerge folder.