Relax timing requirement of one unit test slightly
One test failed on the build agent because it checked for DateModified
to be within 1 second of "now", and the build agent was running just
slow enough for the DateModified to be 1.05 seconds from "now". I'll
update the test to look for "within 5 seconds", which should be generous
enough to allow for majoy build-agent slowdown and still pass.