Fix test data repos
We need to keep the Windows line endings in the Mercurial repos
because Flexbridge creates the XML files with \r\n line endings.
Previously we always got a dirty working directory in Mercurial
when running the tests on Linux. This resulted in local modifications
and so our tests were not testing what we wanted them to test.