Fix missing dispose for FdoCache
In the synchronize action we cleared the lfproject cache while we
still kept on to the actual lfproject. When we later tried to
clear that cache before ending the application the cache was empty
and so couldn't dispose the FdoCache that got created after syncing.

Instead of clearing the lfproject cache we now specifically dispose
the FW project before syncing.