Increase maximum Mongo connection pool
When importing a project with lots of entries (over 3,500 entries in the
case of the test project that triggered this behavior for me), our code
can sometimes run fast enough that it uses up all the connections in the
Mongo driver's default connection pool (default 100 connections). Upping
the pool size by a factor of ten should be enough to handle any project:
by the time we're trying to use the 1,000th connection, the oldest
connections have been torn down and are ready for re-use.