Edit conflicts now expire LF cache correctly
When an item is edited "over there" as well as "here" during a merge,
the DateModified of the conflicted entry's Mongo record will be set to
the current time during the merge. This ensures that it will be newer
than the DateModified of the value in any LF user's browser cache, so
that users will see the new value after the merge.

NOTE: This is the DateModified in Mongo, not the one in FDO. FDO's
DateModified field on entries corresponds to AuthorInfo.ModifiedDate in
the Mongo database. That field gets set by Chorus, and we don't touch it.