Work around buggy implementation of FDO function
The WritingSystemServices.GetWritingSystemList() function that I was
relying on in FDO doesn't handle single-writing-system fields correctly
in FW 8.3; it instead always treats them as being analysis fields, even
if they were really vernacular fields. There's a commit from 2014 that
fixes the problem (commit c371150e by Jason Naylor), but that commit was
never pulled into the 8.3 branches so we don't have it.

In the long run, we need to get that fix into FW 8.3. But in the short
run, we can work around the bug easily enough.