Properly delete FDO fields that were deleted in LF
Pronunciation, lexeme, and etymology fields in FDO have their own
objects, which were not being deleted if an LF user deleted the relevant
fields. For cases like pronunciation which have multiple related fields,
we only delete the FDO object if *all* relevant LF fields were deleted
(that is, the user set them to empty in LF). That is, if all four of
Pronunciation, CV Pattern, Tone and Location are deleted in LF, then
(and only then) will the FDO Pronunciation object be deleted.