Have rsyslog create Trello cards when LfMerge fails
If LfMerge throws an exception and as a result ends up putting a project
on hold, rsyslog will send an email to the Trello card-creation address
to create a new Trello card with the log message (which contains the
exception text).

This has already been run on the live server.