Actually fix MongoWaitQueueFullExceptions
My previous fix had a bug, and I wasn't actually caching the connections
that I thought I was caching. This fixes that bug, and now the database
connections are actually cached.

If we still see MongoWaitQueueFullExceptions despite this fix, there's
another fix to try, but we'll try this one first.

(New commit needed because previous commit was too hasty and didn't
actually compile.)