Catch Mongo wait queue exceptions and retry 5 times
This is potentially dangerous, depending on how the Mongo driver handles
wait queue exceptions. If, when Mongo throws a wait queue exception, it
has given up on the operation, then this code will be fine. But if
despite throwing the exception, Mongo is still planning to apply the
operation at some point, then non-idempotent operations like inserts
could be applied twice.

Therefore, if we see new records showing up duplicated, we should
re-examine this code and maybe revert it.