WIP: Cache writing system info from FDO
The unit tests have NOT yet been run on this code; I started having a
problem with my lfmerge-fdo package right around the time I committed
this and started trying to run the tests. This commit is so that others
can pull it and try running the tests, while I fix my lfmerge-fdo install.