Skip synchronize after initial clone
We use a new project state called CLONED to indicate that an initial
clone (and initial TransferFdoToMongo action) has succeeded, so no
further action is necessary. When LfMerge exits, a CLONED state will
become IDLE, so that the *next* time LfMerge is run, it will perform a
Synchronize action as normal.