Fix crash when adding new senses in FW after S/R
If a new entry was created in LanguageForge, then a Send/Receive was
done so that the new entry ended up in FieldWorks, you couldn't add new
senses to that entry, or FieldWorks would throw a NullReferenceException.
It turns out that we weren't setting the lexeme form correctly in
LfMerge; specifically, we were forgetting to set the lexeme form's
MorphTypeRA field, and there is some FW code that assumes that the
MorphTypeRA will be set and doesn't handle null morph types correctly.
This commit fixes our mistake, so we now set morph types correctly in
new entries created in LF.