When state file doesn't exist, it is not an error
If we log this as an error, it will create a card in our Trello board.
That's not desired for this particular situation, which is usually not
an error (it usually happens when someone clones a project for the first
time). So I'm downgrading the severity to Notice, which won't create a
Trello card from the log message.