Add integration test for LfMergeBridge conflict scenario
This change adds an integration test for the scenario described in
[Fix LfMergeBridge conflict scenario](https://trello.com/c/UoyXj4jP)
where we didn't pick up changes made in FLEx if we made changes to
different fields in the same entry.

The scenario will be fixed with [flexbridge's
PR#137](https://github.com/sillsdev/flexbridge/pull/137).

Note that this change modifies the logic of one test: previously we
expected DateModified to stay the same if both changed the same
data. However, that is a merge conflict and flexbridge will now
update the timestamp for merge conflicts, so the test now expects
DateModified to be greater than the previous value.