Fix custom field config disappearing
We had been building up the custom field config at the same time we were
getting data from custom fields, but once we switched to only processing
entries with changes, that approach became the wrong thing to do. It was
only creating config if an entry with custom field data happened to be
changed in this S/R.

Now, instead of calculating custom field config for every entry, we
calculate it *once*, before the entry-processing loop. This speeds up
our S/R cycle, as a bonus. In addition, we also use the service locator
cache (previously introduced for FdoEntryToLfEntry) in handling custom
field data, for an extra speed boost.