Workaround for migration curiosity
When doing the data migration from 7000069 to 7000070 the
abbreviation field for the English writing system gets changed
from "en" to "Eng" although the abbreviations in other ws don't
get changed. This change makes the same change in our mongo
test double so that we don't introduce an extra hg commit which
our tests don't expect.