Don't skip Sync step if webwork folder is missing
There's a scenario which could result in loss of user data that had been
stored in Mongo:

1. User makes changes in LF.
2. While he's making those changes, the project's webwork folder
   (containing the Mercurial repo) gets deleted somehow.
3. User pushes Send/Receive button.
4. LfMerge sees no Mercurial repo, and thinks this is an initial clone.
5. Because this is an initial clone, TransferMongoToFdo is skipped.
6. Result: the TransferFdoToMongo step overwrites the user's changes in
   LF, which are permanently lost since Mongo doesn't keep previous
   versions of overwritten data.

To avoid this scenario, if the project's Mercurial repo is missing, we
also check whether the Mongo database contains a project record for this
project. If that record exists and contains a known piece of data (a
Send/Receive project identifier, which should exist for every such
project), then it's a real project (as opposed to a "ghost" left behind
from project deletion) and we should NOT skip the Mongo->Fdo step, lest
user data from LF be lost.