Fix mercurial.ini file to include all expected extensions
Chorus tries to update the mercurial.ini file on the fly at runtime.
However if the mercurial.ini got installed by the package we don't
have write permissions and so this fails. This change adds the
eol extension that got recently added for Linux in Chorus, and
also rewrites the entire file so that the we use \n line endings
(which is also a cosmetic improvement :-)).