Prevent crash when custom list fields are empty
The FDO code for get_ObjectProp unconditionally casts the received value
to int, but for vector properties it's received an IEnumerable, not an
int. This cast causes a crash, which is a problem. Since there is
therefore no way to check ahead of time if the field was already null,
we have to simply set it to null unconditionally. This may have an
undesirable side-effect of causing the modified date of the field to be
updated, unfortunately.