Fix bug where only one custom list item was added
If a ReferenceCollection (or ReferenceSequence) custom field had more
than one item added at once in LF, only the first item was being added
by LfMerge. The ultimate cause of this is that the GUIDs weren't being
set properly in LF, but this workaround in LfMerge also fixes the bug.