Fix bug with custom MultiOptionList fields
Custom MultiOptionList fields, when edited in LF, weren't getting the
edits properly showing up in FW. That's because LfMerge had basically
assumed that LF would update the customFieldGuids field in Mongo when
users edited MultiOptionLists (adding or removing items), but LF wasn't
dealing with GUIDs at all. So I rewrote the relevant part of LfMerge's
custom field handling to expect LF to only hand it strings (representing
the unique keys of CmPossibility objects), and wrote some unit tests to
verify that the bug is fixed.