Fix Thai input system getting into non-Thai projects
The cause of the problem was a combination of two things:

1. The PHP code that creates a LexProjectModel creates two input
systems, 'en' and 'th', as "placeholders" so that its config can be
correctly updated.

2. The "placeholder" input systems should normally be overwritten on
initial clone. But when LfMerge detects an older project model and calls
itself recursively, the "inner" LfMerge process sees an existing Mongo
project and thinks that it contains user data that must be preserved, so
it doesn't treat it as an initial clone. (This was a recent change
intended to preserve user data if the webwork/project-code folder ever
gets corrupted or deleted.) Then because it thinks there is valuable
user data in Mongo, it does a TransferMongoToFdo first before doing the
TransferFdoToMongo, and so the placeholder 'th' input system is added to
the FDO project.

The solution I've implemented is to add a command-line parameter to
LfMerge, which the "outer" LfMerge call can use to determine that this
is actually an initial clone. If the --initial-clone parameter is
passed, LfMerge will overwrite the existing Mongo config (including the
placeholder input systems), and will skip the TransferMongoToFdo step,
thus throwing away the data currently in Mongo because it's not actually
real user data. (Obviously, using the --initial-clone parameter when
there *is* valuable user data in Mongo would be a Bad Thing™, since that
valuable user data could be lost.)