Fix Thai input system getting into non-Thai projects
Fixes https://github.com/sillsdev/web-languageforge/issues/101

The cause of the problem was a combination of two things:

1. The PHP code that creates a LexProjectModel creates two input
systems, 'en' and 'th', as "placeholders" so that its config can be
correctly updated.

2. The "placeholder" input systems should normally be overwritten on
initial clone. But when LfMerge detects an older project model and calls
itself recursively, the "inner" LfMerge process sees an existing Mongo
project and thinks that it contains user data that must be preserved, so
it doesn't treat it as an initial clone. (This was a recent change
intended to preserve user data if the webwork/project-code folder ever
gets corrupted or deleted.) Then because it thinks there is valuable
user data in Mongo, it does a TransferMongoToFdo first before doing the
TransferFdoToMongo, and so the placeholder 'th' input system is added to
the FDO project.

The solution I've implemented is to check that the project exists and
has a LastSyncedDate more recent than the Unix epoch (1970-01-01). If
those two conditions apply, this will *not* be an initial clone and so
the TransferFdoToMongo step will run. Alternately, if the project has at
least one lexical entry in the Mongo DB, we also do not consider this an
initial clone and thus run TransferFdoToMongo.