WIP on comment-converting code
S/R of comments works now, but LF comments are being duplicated by the
S/R process, so it's not ready to merge yet. It is, however, in a good
enough state to commit and save the work done so far.

Note that when we merge this into master, we also need to merge the
corresponding FlexBridge branch into its `lfmerge` branch. Both of those
merges need to happen at the same time, since they rely on each other.